+++ test/endianness/SequenceEndianness.cpp
@@ -59,7 +59,7 @@ TEST_P(SequenceEndianness, Char)
 TEST_P(SequenceEndianness, Int8)
 {
     int8_t input[ARRAY_CAPACITY];
-    std::fill_n(input, SEQUENCE_SIZE, static_cast<int8_t>(0x09));
+    std::fill_n(input, SEQUENCE_SIZE, int8_t(0x09));
     int8_t output[ARRAY_CAPACITY];
 
     EXPECT_TRUE(mc_serialize_endian_sequence_int8_t(&writer, endianness, input, SEQUENCE_SIZE));
@@ -71,7 +71,7 @@ TEST_P(SequenceEndianness, Int8)
 TEST_P(SequenceEndianness, Uint8)
 {
     uint8_t input[ARRAY_CAPACITY];
-    std::fill_n(input, SEQUENCE_SIZE, static_cast<uint8_t>(0x09));
+    std::fill_n(input, SEQUENCE_SIZE, uint8_t(0x09));
     uint8_t output[ARRAY_CAPACITY];
 
     EXPECT_TRUE(mc_serialize_endian_sequence_uint8_t(&writer, endianness, input, SEQUENCE_SIZE));
@@ -83,7 +83,7 @@ TEST_P(SequenceEndianness, Uint8)
 TEST_P(SequenceEndianness, Int16)
 {
     int16_t input[ARRAY_CAPACITY];
-    std::fill_n(input, SEQUENCE_SIZE, static_cast<int16_t>(0x0A0B));
+    std::fill_n(input, SEQUENCE_SIZE, int16_t(0x0A0B));
     int16_t output[ARRAY_CAPACITY];
 
     EXPECT_TRUE(mc_serialize_endian_sequence_int16_t(&writer, endianness, input, SEQUENCE_SIZE));
@@ -95,7 +95,7 @@ TEST_P(SequenceEndianness, Int16)
 TEST_P(SequenceEndianness, Uint16)
 {
     uint16_t input[ARRAY_CAPACITY];
-    std::fill_n(input, SEQUENCE_SIZE, static_cast<uint8_t>(0x0A0B));
+    std::fill_n(input, SEQUENCE_SIZE, uint16_t(0x0A0B));
     uint16_t output[ARRAY_CAPACITY];
 
     EXPECT_TRUE(mc_serialize_endian_sequence_uint16_t(&writer, endianness, input, SEQUENCE_SIZE));
